requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pandas>=2.0.3
python-dotenv>=1.0.0
openai>=0.28.0
chromadb>=0.4.13
pydantic>=2.3.0
